+++ src/consts.rs
@@ -1,7 +1,10 @@
 pub const TMP_DIR: &str = "/tmp/firesquid";
 pub const LOG_DIR: &str = "/tmp/firesquid/logs";
 
-pub const ASSETS_DIR: &str = "/etc/firesquid";
+pub const SOCKET: &str = "/tmp/firesquid/http.sock";
+
+// pub const ASSETS_DIR: &str = "/etc/firesquid";
+pub const ASSETS_DIR: &str = "./assets";
 
 pub const DRIVE_NAME: &str = "rootfs";
 pub const KERNEL_NAME: &str = "vmlinux";
